skills/mr-perfection/dap-chat/pyproject.toml
[project]
name = "dap-openclaw"
version = "0.1.0"
description = "OpenClaw skill that gives any agent DAP capabilities"
requires-python = ">=3.11"
dependencies = [
    "dap-core",
    "dap-relay",
    "dap-registry",
    "httpx>=0.27.0",
    "websockets>=13.0",
    "pydantic>=2.0.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "uvicorn>=0.34.0",
]

[tool.hatch.build.targets.wheel]
packages = ["dap_skill"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
dap-core = { path = "../../packages/dap-core", editable = true }
dap-relay = { path = "../../packages/dap-relay", editable = true }
dap-registry = { path = "../../packages/dap-registry", editable = true }
